refractor(): Move files Error.tsx
